// src/app/modules/dashboard/principal/models/convocado.ts
export class Convocado{
    codigo: number = 0;
    dniRuc: string = '';
    nombre: string = '';
    apellidoPaterno: string = '';
    apellidoMaterno: string = '';
    direccion: string = '';
    correo: string = '';
    telefono: string = '';
    urlPDF: string = '';
    estado: string= '';
    codigoContratacion: number = 0;
}